requirements.txt
pyserial
dweepy